[Refactoring]: split devices.noise_utils into qis.noise_utils and devices.noise_utils
