+++ cirq-core/cirq/devices/__init__.py
@@ -46,13 +46,4 @@ from cirq.devices.superconducting_qubits_noise_properties import (
     SuperconductingQubitsNoiseProperties,
 )
 
-from cirq.devices.noise_utils import (
-    OpIdentifier,
-    decay_constant_to_xeb_fidelity,
-    decay_constant_to_pauli_error,
-    pauli_error_to_decay_constant,
-    xeb_fidelity_to_decay_constant,
-    pauli_error_from_t1,
-    average_error,
-    decoherence_pauli_error,
-)
+from cirq.devices.noise_utils import OpIdentifier
