+++ cirq-core/cirq/devices/superconducting_qubits_noise_properties.py
@@ -20,7 +20,7 @@ from dataclasses import dataclass, field
 from functools import cached_property
 from typing import Dict, TYPE_CHECKING, List, Set, Type
 
-from cirq import ops, devices
+from cirq import ops, devices, qis
 from cirq.devices import noise_utils
 
 if TYPE_CHECKING:
@@ -129,7 +129,7 @@ class SuperconductingQubitsNoiseProperties(devices.NoiseProperties, abc.ABC):
     def _get_pauli_error(self, p_error: float, op_id: noise_utils.OpIdentifier):
         time_ns = float(self.gate_times_ns[op_id.gate_type])
         for q in op_id.qubits:
-            p_error -= noise_utils.decoherence_pauli_error(self.t1_ns[q], self.tphi_ns[q], time_ns)
+            p_error -= qis.decoherence_pauli_error(self.t1_ns[q], self.tphi_ns[q], time_ns)
         return p_error
 
     @cached_property
