+++ cirq-core/cirq/qis/__init__.py
@@ -50,3 +50,13 @@ from cirq.qis.states import (
     validate_qid_shape,
     validate_normalized_state_vector,
 )
+
+from cirq.qis.noise_utils import (
+    decay_constant_to_xeb_fidelity,
+    decay_constant_to_pauli_error,
+    pauli_error_to_decay_constant,
+    xeb_fidelity_to_decay_constant,
+    pauli_error_from_t1,
+    average_error,
+    decoherence_pauli_error,
+)
